Adding Accessibility Capability
Accessibility is an important part of modern development. The package added allows us to run against the command line (using the YAML file) and gives us an option to include this programmatically in test cases if we so wish. The accompanying JS file is an illustrative example taken from https://openbase.com/js/accessibility-checker but is functional enough to show how to use it. Given this is a mono repo there might be a better location to land the JS file so leaving it in root for this review