Adding tssl support
TSSL is a spec driven development tool but can be used by any AI agent. The intent is to create tiles to allow any Agentic Code generator to "understand" better the libraries being used by the application. Effectively, it gives a pattern and an understanding that allows the agent track changes to upstream dependencies and in essence provide better guided usage of AI in Tutors